+++ sudokuru/app/Components/Sudoku Board/Functions/generateGame.ts
@@ -5,7 +5,7 @@ import {List} from "immutable";
 
 
 // Sudokuru Package Import
-const sudokuru = require("../../../node_modules/sudokuru/dist/bundle.js");
+const sudokuru = require("../../../../node_modules/sudokuru/dist/bundle.js");
 
 // Sudokuru Package Constants
 const Puzzles = sudokuru.Puzzles;
